pyproject.toml
[project]
name = "dna-sample-manager"
version = "0.1.0"
description = "Application web de gestion d'échantillons ADN"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.1",
    "openpyxl>=3.1.0",
    "typer>=0.12.0",
]

[project.scripts]
dna-sample-manager = "dna_sample_manager.cli:app"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
